pyproject.toml
[project]
name = "ezCaptions"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "assemblyai>=0.52.4",
    "fastapi>=0.134.0",
    "google-cloud-storage>=2.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "python-semantic-release>=9.0.0",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["app/__init__.py:__version__"]
commit_message = "🔖 v{version}"
commit_parser = "emoji"
major_on_zero = false
allow_zero_version = true

[tool.semantic_release.commit_parser_options]
major_tags = ["💥"]
minor_tags = ["✨", "🚑", "🎉"]
patch_tags = ["🐛", "🩹", "🔒"]

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
